Recent Activations fieldset: extract to shared lib, add to Team Edit
The "Recent myAvailability Activation Requests" fieldset (expand a recent
unit activation, click an Accepted/Available/Conditional responder
straight into the team) lived inline in the Team Create inject script.

- move it to src/lib/teamRecentActivations.js, exporting
  initRecentActivationsFieldset(vm) + addTeamMemberById(vm, memberId),
  parameterised on the team-form viewmodel (Create exposes it as
  `teamViewModel`, Edit as `vm`)
- wire it into the Team Edit inject script
- rename styles/teams.create.css -> teams.recent-activations.css (content
  unchanged) and load it from both the create and edit content scripts

No behaviour change on Team Create; Team Edit gains the fieldset. The
train/dev placeholder and prod-only gating (#422) carry over unchanged.
