+++ src/contentscripts/teams/create.js
@@ -1,5 +1,5 @@
 var inject = require('../../lib/inject.js');
-require('../../styles/teams.create.css');
+require('../../styles/teams.recent-activations.css');
 
 //inject our JS resource
 inject('teams/create.js');
