+++ src/contentscripts/teams/edit.js
@@ -1,4 +1,5 @@
 var inject = require('../../lib/inject.js');
+require('../../styles/teams.recent-activations.css');
 
 //inject our JS resource
 inject('teams/edit.js');
